Replace select and move util
