+++ packages/manager/src/components/PaginationFooter/PaginationFooter.constants.ts
@@ -1,0 +1,3 @@
+export const MIN_PAGE_SIZE = 25;
+
+export const PAGE_SIZES = [MIN_PAGE_SIZE, 50, 75, 100, Infinity];
